requirements.txt
google-generativeai>=0.8.0
selenium>=4.0.0
webdriver-manager>=4.0.0
pyautogui>=0.9.0
beautifulsoup4>=4.0.0
elevenlabs>=0.2.0
pyttsx3>=2.90
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.0.0